+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `datacontract test --output-format junit --output TEST-datacontract.xml` Export CLI test results
   to a file, in a standard format (e.g. JUnit) to improve CI/CD experience (#650)
 
+### Changed
+
+- Changed Docker base image to python:3.11-bullseye
+
 ## [0.10.22] - 2025-02-20
 
 ### Added