+++ CHANGELOG.md
@@ -11,7 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `datacontract test --output-format junit --output TEST-datacontract.xml` Export CLI test results
   to a file, in a standard format (e.g. JUnit) to improve CI/CD experience (#650)
+
 - Added import for `ProtoBuf`
+
+- `dbt` & `dbt-sources` export formats now support the optional `--server` flag to adapt the DBT column `data_type` to specific SQL dialects
+
 
 ### Changed
 
@@ -21,6 +25,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Unicode Encode Error when exporting data contract YAML to HTML
   (#652)
+- Fix multiline descriptions in the DBT export functionality
+- Incorrectly parsing $ref values in definitions (#664)
+- Better error message when the server configuration is missing in a data contract (#670)
+- Improved default values in ODCS generator to avoid breaking schema validation (#671)
+- Updated ODCS v3 generator to drop the "is" prefix from fields like `isNullable` and `isUnique` (#669)
 
 ## [0.10.22] - 2025-02-20