+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Changed Docker base image to python:3.11-bullseye
+- Relax fastparquet dependency
 
 ### Fixed
 
@@ -26,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Better error message when the server configuration is missing in a data contract (#670)
 - Improved default values in ODCS generator to avoid breaking schema validation (#671)
 - Updated ODCS v3 generator to drop the "is" prefix from fields like `isNullable` and `isUnique` (#669)
+- Fix issue when testing databricks server with ODCS format
 
 ## [0.10.22] - 2025-02-20